+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_encoder.rs
@@ -258,7 +258,7 @@ impl ArrayPartialEncoderTraits for ShardingPartialEncoder {
         // Read the straddling inner chunks
         let inner_chunks_encoded = self
             .input_output_handle
-            .partial_decode(&mut byte_ranges.into_iter(), options)?
+            .partial_decode(Box::new(byte_ranges.into_iter()), options)?
             .map(|bytes| bytes.into_iter().map(Cow::into_owned).collect::<Vec<_>>());
 
         // Decode the straddling inner chunks