+++ zarrs/src/array/codec/byte_interval_partial_decoder.rs
@@ -41,10 +41,10 @@ impl BytesPartialDecoderTraits for ByteIntervalPartialDecoder {
 
     fn partial_decode(
         &self,
-        byte_ranges: &mut dyn ByteRangeIterator,
+        byte_ranges: ByteRangeIterator,
         options: &CodecOptions,
     ) -> Result<Option<Vec<RawBytes<'_>>>, CodecError> {
-        let mut byte_ranges = byte_ranges.map(|byte_range| match byte_range {
+        let byte_ranges = byte_ranges.map(|byte_range| match byte_range {
             ByteRange::FromStart(offset, None) => {
                 ByteRange::FromStart(self.byte_offset + offset, Some(self.byte_length))
             }
@@ -55,7 +55,8 @@ impl BytesPartialDecoderTraits for ByteIntervalPartialDecoder {
                 ByteRange::FromStart(self.byte_offset + self.byte_length - length, Some(length))
             }
         });
-        self.input_handle.partial_decode(&mut byte_ranges, options)
+        self.input_handle
+            .partial_decode(Box::new(byte_ranges), options)
     }
 }
 
@@ -89,12 +90,12 @@ impl AsyncByteIntervalPartialDecoder {
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl AsyncBytesPartialDecoderTraits for AsyncByteIntervalPartialDecoder {
-    async fn partial_decode(
-        &self,
-        byte_ranges: &mut dyn ByteRangeIterator,
+    async fn partial_decode<'a>(
+        &'a self,
+        byte_ranges: ByteRangeIterator<'a>,
         options: &CodecOptions,
     ) -> Result<Option<Vec<RawBytes<'_>>>, CodecError> {
-        let mut byte_ranges = byte_ranges.map(|byte_range| match byte_range {
+        let byte_ranges = byte_ranges.map(|byte_range| match byte_range {
             ByteRange::FromStart(offset, None) => {
                 ByteRange::FromStart(self.byte_offset + offset, Some(self.byte_length))
             }
@@ -106,7 +107,7 @@ impl AsyncBytesPartialDecoderTraits for AsyncByteIntervalPartialDecoder {
             }
         });
         self.input_handle
-            .partial_decode(&mut byte_ranges, options)
+            .partial_decode(Box::new(byte_ranges), options)
             .await
     }
 }