+++ zarrs/src/array/codec/bytes_to_bytes/adler32.rs
@@ -154,7 +154,10 @@ mod tests {
                 .unwrap();
             assert_eq!(partial_decoder.size(), input_handle.size()); // adler32 partial decoder does not hold bytes
             let decoded_partial_chunk = partial_decoder
-                .partial_decode(&mut decoded_regions.into_iter(), &CodecOptions::default())
+                .partial_decode(
+                    Box::new(decoded_regions.into_iter()),
+                    &CodecOptions::default(),
+                )
                 .unwrap()
                 .unwrap();
             let answer: &[Vec<u8>] = &[vec![3, 4]];
@@ -195,7 +198,10 @@ mod tests {
                 .await
                 .unwrap();
             let decoded_partial_chunk = partial_decoder
-                .partial_decode(&mut decoded_regions.into_iter(), &CodecOptions::default())
+                .partial_decode(
+                    Box::new(decoded_regions.into_iter()),
+                    &CodecOptions::default(),
+                )
                 .await
                 .unwrap()
                 .unwrap();