+++ zarrs/src/array/codec/bytes_to_bytes/strip_suffix_partial_decoder.rs
@@ -37,14 +37,14 @@ impl BytesPartialDecoderTraits for StripSuffixPartialDecoder {
 
     fn partial_decode(
         &self,
-        decoded_regions: &mut dyn ByteRangeIterator,
+        decoded_regions: ByteRangeIterator,
         options: &CodecOptions,
     ) -> Result<Option<Vec<RawBytes<'_>>>, CodecError> {
         decoded_regions
             .map(|decoded_region| {
                 let bytes = self
                     .input_handle
-                    .partial_decode_concat(&mut [decoded_region].into_iter(), options)?;
+                    .partial_decode_concat(Box::new([decoded_region].into_iter()), options)?;
                 Ok::<_, CodecError>(bytes.map(|bytes| match decoded_region {
                     ByteRange::FromStart(_, Some(_)) => bytes,
                     ByteRange::FromStart(_, None) => {
@@ -87,9 +87,9 @@ impl AsyncStripSuffixPartialDecoder {
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl AsyncBytesPartialDecoderTraits for AsyncStripSuffixPartialDecoder {
-    async fn partial_decode(
-        &self,
-        decoded_regions: &mut dyn ByteRangeIterator,
+    async fn partial_decode<'a>(
+        &'a self,
+        decoded_regions: ByteRangeIterator<'a>,
         options: &CodecOptions,
     ) -> Result<Option<Vec<RawBytes<'_>>>, CodecError> {
         use futures::{StreamExt, TryStreamExt};
@@ -98,13 +98,13 @@ impl AsyncBytesPartialDecoderTraits for AsyncStripSuffixPartialDecoder {
             match decoded_region {
                 ByteRange::FromStart(_, Some(_)) => Ok::<_, CodecError>(
                     self.input_handle
-                        .partial_decode_concat(&mut [decoded_region].into_iter(), options)
+                        .partial_decode_concat(Box::new([decoded_region].into_iter()), options)
                         .await?,
                 ),
                 ByteRange::FromStart(_, None) | ByteRange::Suffix(_) => {
                     let bytes = self
                         .input_handle
-                        .partial_decode_concat(&mut [decoded_region].into_iter(), options)
+                        .partial_decode_concat(Box::new([decoded_region].into_iter()), options)
                         .await?;
                     if let Some(bytes) = bytes {
                         let length = bytes.len() - self.suffix_size;